refactor(jsi): move the QuickJS engine layer into the shared tree
First engine through the shared-layer split, chosen as the pilot because it
needed the least surgery: its contract header had two Objective-C references
and its three implementation files had none.

The `nativescript::engine` classes wrapping QuickJS -- Runtime, Value, Object,
Function, Array, String, BigInt, ArrayBuffer and the host-object plumbing --
move to jsi/quickjs/ and become .cpp. They were already plain C++; the .mm
extension was incidental, inherited from living under ffi/objc.

The Apple bridge's baggage stays behind in ffi/objc/quickjs/, which now holds
Foundation, <objc/*>, the Mach-O metadata section lookup, the metadata reader
and the NativeApiClassBuilderProtocol forward declaration, then includes the
neutral header. Apple sources include the same path as before, so no Apple
include changed. The header's other platform includes (dispatch, dlfcn, ffi)
turned out to be vestigial -- nothing in the header body used them.

Proof this is genuinely neutral, rather than merely compiling on a Mac: the
moved header passes -fsyntax-only as plain C++ with no Foundation in reach.
That matters because #import <Foundation/Foundation.h> compiles fine under
-x c++ on macOS, so a green Apple build alone would prove nothing.

Verified: iOS TestRunner on QuickJS, 713 specs, 0 failures.
check_jsi_layer_neutral.sh passes. No Android file touched.

(cherry picked from commit cf98d6ff6875ad83d5c54c177ed5bda94531cfd5)
