+++ NativeScript/CMakeLists.txt
@@ -290,10 +290,12 @@ set(FFI_JSC_ENGINE_SOURCE_FILES
 
 set(FFI_QUICKJS_ENGINE_SOURCE_FILES
   ${FFI_ENGINE_SHARED_SOURCE_FILES}
-  ffi/objc/quickjs/NativeApiQuickJSHostObjects.mm
+  # Platform-neutral engine layer, shared with the Android runtime.
+  jsi/quickjs/QuickJSHostObjects.cpp
+  jsi/quickjs/QuickJSRuntime.cpp
+  jsi/quickjs/QuickJSValue.cpp
+  # Apple-only ObjC bridge on top of it.
   ffi/objc/quickjs/NativeApiQuickJS.mm
-  ffi/objc/quickjs/NativeApiQuickJSRuntime.mm
-  ffi/objc/quickjs/NativeApiQuickJSValue.mm
 )
 
 set(SOURCE_FILES
