+++ NativeScript/ffi/objc/shared/RuntimeCleanupRegistry.h
@@ -1,39 +1,6 @@
-#ifndef NATIVESCRIPT_FFI_OBJC_RUNTIME_CLEANUP_REGISTRY_H
-#define NATIVESCRIPT_FFI_OBJC_RUNTIME_CLEANUP_REGISTRY_H
+#ifndef NATIVESCRIPT_FFI_OBJC_RUNTIME_CLEANUP_REGISTRY_FORWARD_H
+#define NATIVESCRIPT_FFI_OBJC_RUNTIME_CLEANUP_REGISTRY_FORWARD_H
 
-#include <unordered_map>
-#include <utility>
+#include "jsi/shared/RuntimeCleanupRegistry.h"
 
-namespace nativescript::engine {
-
-class RuntimeCleanupRegistry {
- public:
-  using Cleanup = void (*)(void*);
-
-  void track(void* pointer, Cleanup cleanup) { actions_[pointer] = cleanup; }
-
-  template <typename T>
-  void track(T* pointer) {
-    track(pointer, [](void* value) { delete static_cast<T*>(value); });
-  }
-
-  void untrack(void* pointer) { actions_.erase(pointer); }
-  bool empty() const { return actions_.empty(); }
-
-  void cleanup() {
-    while (!actions_.empty()) {
-      auto action = actions_.begin();
-      void* pointer = action->first;
-      Cleanup cleanup = action->second;
-      actions_.erase(action);
-      cleanup(pointer);
-    }
-  }
-
- private:
-  std::unordered_map<void*, Cleanup> actions_;
-};
-
-}  // namespace nativescript::engine
-
-#endif  // NATIVESCRIPT_FFI_OBJC_RUNTIME_CLEANUP_REGISTRY_H
+#endif  // NATIVESCRIPT_FFI_OBJC_RUNTIME_CLEANUP_REGISTRY_FORWARD_H
