+++ NativeScript/jsi/shared/NativeApiStackValueArray.h
